refactor: 로그인, 회원가입 로직 추가 리팩토링 useAuth
